+++ cmd/stakercli/daemon/daemoncommands.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"net/url"
 
+	"github.com/babylonlabs-io/btc-staker/cmd"
 	"github.com/babylonlabs-io/btc-staker/cmd/stakercli/helpers"
 	dc "github.com/babylonlabs-io/btc-staker/stakerservice/client"
 	"github.com/urfave/cli"
@@ -263,7 +265,7 @@ var withdrawableTransactionsCmd = cli.Command{
 // checkHealth checks if staker daemon is running.
 func checkHealth(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -284,7 +286,7 @@ func checkHealth(ctx *cli.Context) error {
 // listOutputs lists current unspent outputs in connected wallet.
 func listOutputs(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -305,7 +307,7 @@ func listOutputs(ctx *cli.Context) error {
 // babylonFinalityProviders lists current finality providers.
 func babylonFinalityProviders(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -338,7 +340,7 @@ func babylonFinalityProviders(ctx *cli.Context) error {
 // stake stakes a BTC.
 func stake(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -363,7 +365,7 @@ func stake(ctx *cli.Context) error {
 // stakeFromPhase1TxBTC delegates a staking transaction from a phase 1 tx BTC.
 func stakeFromPhase1TxBTC(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -400,7 +402,7 @@ func stakeFromPhase1TxBTC(ctx *cli.Context) error {
 
 func unstake(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -422,7 +424,7 @@ func unstake(ctx *cli.Context) error {
 // unbondStaking unbonds a staking transaction.
 func unbond(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -444,7 +446,7 @@ func unbond(ctx *cli.Context) error {
 // stakingDetails gets the details of a staking transaction.
 func stakingDetails(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -466,7 +468,7 @@ func stakingDetails(ctx *cli.Context) error {
 // listStakingTransactions lists all the staking transactions.
 func listStakingTransactions(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return fmt.Errorf("failed to create staker service JSON-RPC client: %w", err)
 	}
@@ -499,7 +501,7 @@ func listStakingTransactions(ctx *cli.Context) error {
 // withdrawableTransactions lists all the withdrawable staking transactions.
 func withdrawableTransactions(ctx *cli.Context) error {
 	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
-	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
+	client, err := NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
 	}
@@ -525,4 +527,21 @@ func withdrawableTransactions(ctx *cli.Context) error {
 	helpers.PrintRespJSON(transactions)
 
 	return nil
+}
+
+// NewStakerServiceJSONRPCClient creates a client connection with basic auth
+// The username and password are loaded from environment variables
+func NewStakerServiceJSONRPCClient(remoteAddressWithoutAuth string) (*dc.StakerServiceJSONRPCClient, error) {
+	parsedURL, err := url.Parse(remoteAddressWithoutAuth)
+	if err != nil {
+		return nil, err
+	}
+
+	user, pwd, err := cmd.GetEnvBasicAuth()
+	if err != nil {
+		return nil, err
+	}
+	parsedURL.User = url.UserPassword(user, pwd)
+
+	return dc.NewStakerServiceJSONRPCClient(parsedURL.String())
 }